Before your eyes, Joaquin is born.
The obvious question arises immediately. WTF? Why fork Hugo, rename
it, make fun of the newly appropriated name?

Well, as the well-known poet well-put it, the jokes on you. As of
me, my reasoning for this project go well beyond a plain humorous
ornament. I appreciate the legacy of Steven Francia as much as the
next man; in fact, I use Hugo on the daily myself and I must say
it served well for me over the course of last couple years. I am
talking of https://aletheia.icu/~badt, my home page and many more
to come.

I can't yet say much about Aletheia, the project, apart from the
obvious fact that it currently hosts my blog and will soon enough
host blogs of many more people to come. Hugo in its present form,
unfortunately isn't equipped with the right set of tools and it's
rather opionated, of course, which can be a good thing, or it might
as well be a bad thing. In case of Aletheia, it's a bad thing.

Hence, Joaquin is born.